# Conflicts:
#	src/main/resources/application.properties

+++ src/main/java/com/bigeng/invoicing/service/base/SaleOrderDetailService.java
@@ -12,7 +12,7 @@ public interface SaleOrderDetailService {
 
     SaleOrderDetail selectByPrimaryKey(SaleOrderDetailKey key);
 
-    SaleOrderDetail findOrderByNo(Integer iNo);
+    SaleOrderDetail findOrderByNo(Long iNo);
 
     int updateByPrimaryKeySelective(SaleOrderDetail record);
